requirements.txt
flask==3.0.0
flask-limiter==3.5.0
flask-cors==4.0.0
bleach==6.1.0
python-dotenv==1.0.0
anthropic==0.25.0
httpx==0.27.0
gunicorn==21.2.0
